Fix type of master_account variable
master_account variable was defined as int. By defining the var as int,
would cause issues when the account id started with '0'. For example an
account id with 0nnnnnnnnnnn, would become nnnnnnnnnnn. This whould fail
on the regex '[0-9]{12}', because the account id would be striped of the
leading 0, and only have 11 digits.
Changing the master_account to str solves the issue.